Creado bean, dto y dao de suscripcion con sus detalles, deje las bases para iniciar con los services y el controller
