+++ src/main/java/com/devs/powerfit/abstracts/AbstractBean.java
@@ -1,12 +1,18 @@
 package com.devs.powerfit.abstracts;
 
-import com.devs.powerfit.interfaces.IBean;
+import com.devs.powerfit.interfaces.bases.IBean;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 
 @MappedSuperclass
 @Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public abstract class AbstractBean implements IBean {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
