+++ src/main/java/com/devs/powerfit/abstracts/AbstractDto.java
@@ -1,6 +1,6 @@
 package com.devs.powerfit.abstracts;
 
-import com.devs.powerfit.interfaces.IDto;
+import com.devs.powerfit.interfaces.bases.IDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public abstract class AbstractDto implements IDto {
     private Long id;
+    private boolean active;
 
     @Override
     public Long getId(){
