+++ src/main/java/com/devs/powerfit/controllers/suscripciones/SuscripcionController.java
@@ -1,0 +1,9 @@
+package com.devs.powerfit.controllers.suscripciones;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/suscripciones")
+public class SuscripcionController {
+}
