+++ src/main/java/com/devs/powerfit/daos/suscripciones/SuscripcionDao.java
@@ -1,0 +1,10 @@
+package com.devs.powerfit.daos.suscripciones;
+
+import com.devs.powerfit.beans.suscripciones.SuscripcionBean;
+import com.devs.powerfit.dtos.suscripciones.SuscripcionDto;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SuscripcionDao extends JpaRepository<SuscripcionBean, SuscripcionDto> {
+}
