+++ src/main/java/com/devs/powerfit/daos/suscripciones/SuscripcionDetalleDao.java
@@ -1,0 +1,8 @@
+package com.devs.powerfit.daos.suscripciones;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SuscripcionDetalleDao extends JpaRepository {
+}
