--- src/main/java/com/devs/powerfit/dtos/ActividadDto.java
@@ -1,12 +1,0 @@
-package com.devs.powerfit.dtos;
-
-import com.devs.powerfit.abstracts.AbstractDto;
-import lombok.Data;
-
-@Data
-public class ActividadDto extends AbstractDto {
-    private String nombre;
-    private String descripcion;
-    private Double costoMensual;
-    private Double costoSemanal;
-}
