--- src/main/java/com/devs/powerfit/dtos/ClienteDto.java
@@ -1,17 +1,0 @@
-package com.devs.powerfit.dtos;
-
-import com.devs.powerfit.abstracts.AbstractDto;
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class ClienteDto extends AbstractDto {
-    private String nombre;
-    private String ruc;
-    private String cedula;
-    private String telefono;
-    private String email;
-    private String direccion;
-    private LocalDate fechaRegistro;
-}
