--- src/main/java/com/devs/powerfit/dtos/ErrorDto.java
@@ -1,11 +1,0 @@
-package com.devs.powerfit.dtos;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ErrorDto {
-    private String code;
-    private String message;
-}
