+++ src/main/java/com/devs/powerfit/dtos/errors/ErrorDto.java
@@ -1,0 +1,11 @@
+package com.devs.powerfit.dtos.errors;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ErrorDto {
+    private String code;
+    private String message;
+}
