+++ src/main/java/com/devs/powerfit/enums/EEstado.java
@@ -1,0 +1,9 @@
+package com.devs.powerfit.enums;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public enum EEstado {
+    PENDIENTE,
+    PAGADO
+}
