--- src/main/java/com/devs/powerfit/interfaces/IBean.java
@@ -1,10 +1,0 @@
-package com.devs.powerfit.interfaces;
-
-import java.io.Serializable;
-
-public interface IBean extends Serializable {
-
-    public Long getId();
-    public void setId(Long id);
-
-}
