--- src/main/java/com/devs/powerfit/interfaces/IDto.java
@@ -1,11 +1,0 @@
-package com.devs.powerfit.interfaces;
-
-import org.springframework.stereotype.Component;
-
-import java.io.Serializable;
-
-public interface IDto extends Serializable {
-    public Long getId();
-    public void setId(Long id);
-}
-
