+++ src/main/java/com/devs/powerfit/interfaces/bases/IBean.java
@@ -1,0 +1,10 @@
+package com.devs.powerfit.interfaces.bases;
+
+import java.io.Serializable;
+
+public interface IBean extends Serializable {
+
+    public Long getId();
+    public void setId(Long id);
+
+}
