+++ src/main/java/com/devs/powerfit/interfaces/suscripciones/ISuscripcionDetalleService.java
@@ -1,0 +1,7 @@
+package com.devs.powerfit.interfaces.suscripciones;
+
+import com.devs.powerfit.dtos.suscripciones.SuscripcionDetalleDto;
+import com.devs.powerfit.interfaces.bases.IService;
+
+public interface ISuscripcionDetalleService extends IService<SuscripcionDetalleDto> {
+}
