+++ src/main/java/com/devs/powerfit/interfaces/suscripciones/ISuscripcionService.java
@@ -1,0 +1,6 @@
+package com.devs.powerfit.interfaces.suscripciones;
+
+import com.devs.powerfit.dtos.suscripciones.SuscripcionDto;
+import com.devs.powerfit.interfaces.bases.IService;
+public interface ISuscripcionService extends IService<SuscripcionDto> {
+}
